pyproject.toml
[tool.poetry]
name = "visiondeep"
version = "0.1.0"
description = ""
authors = ["jihoahn9303 <rediate03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
torch = "1.12.1"
einops = "^0.6.1"
albumentations = "^1.2.1"
torchvision = "0.13.1"
python-dotenv = "^1.0.0"
omegaconf = "^2.3.0"
loguru = "^0.7.0"
wandb = "^0.15.8"
datasets = "^2.14.4"
pytorch-lightning = "^1.7.7"
torchmetrics = "^0.9.3"
torch-optimizer = "^0.3.0"
rich = "12.6.0"
timm = "0.6.11"
hydra-core = "^1.2.0"
hydra-zen = "v0.9.0rc5"
jaxtyping = "^0.2.21"
beartype = "^0.15.0"
flake8 = "^4.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^4.0.1"
black = "^23.7.0"
pytest = "^7.4.0"
imagehash = "^4.3.1"
pre-commit = "^3.3.3"
mypy = "0.982"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
